+++ CHANGELOG.md
@@ -2,7 +2,7 @@ This file contains the changes to the crate since version 0.4.8.
 
 # 0.7.2
 
- - Add `Primes<N>::factorize` function that returns an iterator over the prime factors of the given number.
+ - Add `Primes<N>::prime_factorization` function that returns an iterator over the prime factors of the given number and their multiplicities.
 
 # 0.7.1